Role-based access for transactions

+++ backend/routes/transactionsRoute.js
@@ -96,68 +96,74 @@ router.put("/:id", async (req, res) => {
             return res.status(400).json({message: "Send all required fields"});
         }
 
-        const { id } = req.params; 
-        const result = await Transaction.findByIdAndUpdate(id, req.body);
-        if (!result) {
-            return res.status(404).json({message: "Transaction not found"});
-        }
-        else { 
-            // send response that transaction has been updated
-            res.status(200).json({message: "Transaction updated successfully"});
-
-            // Send transaction edit details to the admin
-            const adminUser = await User.findOne({ access: 3, verified: true, approved: true });
-            if (!adminUser) {
-                return res.status(404).json({ message: "Admin user not found" });
+        // @ts-ignore
+        const level = req.user.access;
+        if (level > 1 ) {
+            const { id } = req.params; 
+            const result = await Transaction.findByIdAndUpdate(id, req.body);
+            if (!result) {
+                return res.status(404).json({message: "Transaction not found"});
             }
-            const adminEmail = adminUser.email;
+            else { 
+                // send response that transaction has been updated
+                res.status(200).json({message: "Transaction updated successfully"});
 
-            const currentTime = new Date().toLocaleString();
-            const emailSubject = `Transaction Edit - ${currentTime}`;
+                // Send transaction edit details to the admin
+                const adminUser = await User.findOne({ access: 3, verified: true, approved: true });
+                if (!adminUser) {
+                    return res.status(404).json({ message: "Admin user not found" });
+                }
+                const adminEmail = adminUser.email;
 
-            const emailHTML = `
-            <div style="font-family: Arial, sans-serif; max-width: 600px; margin: auto; color: #112376;">
-                <h2 style="color: #112376;">Transaction Edit Notification</h2>
-                <p>Dear Admin,</p>
-                <p>This is to inform you that a transaction with C-Series ${req.body.series} has been edited.</p>
-                <p>
-                    <strong>Editor:</strong> ${
-// @ts-ignore
-                    req.user.name} (${req.user.email})
-                </p>
-                <table border="1" cellpadding="10" style="border-collapse: collapse; width: 100%;">
-                    <tr style="background-color: #112376; color: white;">
-                        <th>Date</th>
-                        <th>Branch</th>
-                        <th>Name of Customer</th>
-                        <th>C-Series</th>
-                        <th>OS</th>
-                        <th>C-Invoice</th>
-                        <th>Seller</th>
-                        <th>Assembler</th>
-                        <th>Total</th>
-                        <th>VAT Sale</th>
-                        <th>VAT Amount</th>
-                    </tr>
-                    <tr>
-                        <td>${req.body.date}</td>
-                        <td>${req.body.branch}</td>
-                        <td>${req.body.name}</td>
-                        <td>${req.body.series}</td>
-                        <td>${req.body.os}</td>
-                        <td>${req.body.invoice}</td>
-                        <td>${req.body.seller}</td>
-                        <td>${req.body.assembler}</td>
-                        <td>${req.body.total}</td>
-                        <td>${req.body.vatsale}</td>
-                        <td>${req.body.vatamount}</td>
-                    </tr>
-                </table>
-                <p>This update was made at ${currentTime}.</p>
-                <p>Best regards,<br/>SBNC Transaction System</p>
-            </div>
-            `;
-            await sendAdminEditUpdate(adminEmail, emailSubject, emailHTML);
+                const currentTime = new Date().toLocaleString();
+                const emailSubject = `Transaction Edit - ${currentTime}`;
+
+                const emailHTML = `
+                <div style="font-family: Arial, sans-serif; max-width: 600px; margin: auto; color: #112376;">
+                    <h2 style="color: #112376;">Transaction Edit Notification</h2>
+                    <p>Dear Admin,</p>
+                    <p>This is to inform you that a transaction with C-Series ${req.body.series} has been edited.</p>
+                    <p>
+                        <strong>Editor:</strong> ${
+    // @ts-ignore
+                        req.user.name} (${req.user.email})
+                    </p>
+                    <table border="1" cellpadding="10" style="border-collapse: collapse; width: 100%;">
+                        <tr style="background-color: #112376; color: white;">
+                            <th>Date</th>
+                            <th>Branch</th>
+                            <th>Name of Customer</th>
+                            <th>C-Series</th>
+                            <th>OS</th>
+                            <th>C-Invoice</th>
+                            <th>Seller</th>
+                            <th>Assembler</th>
+                            <th>Total</th>
+                            <th>VAT Sale</th>
+                            <th>VAT Amount</th>
+                        </tr>
+                        <tr>
+                            <td>${req.body.date}</td>
+                            <td>${req.body.branch}</td>
+                            <td>${req.body.name}</td>
+                            <td>${req.body.series}</td>
+                            <td>${req.body.os}</td>
+                            <td>${req.body.invoice}</td>
+                            <td>${req.body.seller}</td>
+                            <td>${req.body.assembler}</td>
+                            <td>${req.body.total}</td>
+                            <td>${req.body.vatsale}</td>
+                            <td>${req.body.vatamount}</td>
+                        </tr>
+                    </table>
+                    <p>This update was made at ${currentTime}.</p>
+                    <p>Best regards,<br/>SBNC Transaction System</p>
+                </div>
+                `;
+                await sendAdminEditUpdate(adminEmail, emailSubject, emailHTML);
+            }
+        } else {
+            return res.status(401).json({message: "Unauthorized Access"});
         }
     } catch (err) {
         console.error(err.message);
@@ -168,13 +174,19 @@ router.put("/:id", async (req, res) => {
 // DELETE
 router.delete("/:id", async (req, res) => {
     try {
-        const { id } = req.params;
-        const result = await Transaction.findByIdAndDelete(id);
-        if (!result) {
-            return res.status(404).json({message: "Transaction not found"});
-        }
-        else {
-            return res.status(200).json({message: "Transaction deleted successfully"});
+        // @ts-ignore
+        const level = req.user.access;
+        if (level > 1) {
+            const { id } = req.params;
+            const result = await Transaction.findByIdAndDelete(id);
+            if (!result) {
+                return res.status(404).json({message: "Transaction not found"});
+            }
+            else {
+                return res.status(200).json({message: "Transaction deleted successfully"});
+            }
+        } else {
+            return res.status(401).json({message: "Unauthorized Access"});
         }
     } catch (err) {
         console.error(err.message);
